REF: restructure the calendar directory (#41)
(cherry picked from commit c167fffedf056c6c77adefac9a79867b578eab78)
