+++ python/rateslib/instruments/bonds/conventions/discounting.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from typing import TYPE_CHECKING, Protocol
 
-from rateslib.calendars import dcf
+from rateslib.scheduling import dcf
 
 if TYPE_CHECKING:
     from rateslib.instruments.bonds.conventions.accrued import AccrualFunction
