+++ python/rateslib/periods/base.py
@@ -4,11 +4,11 @@ from abc import ABCMeta, abstractmethod
 from typing import TYPE_CHECKING
 
 from rateslib import defaults
-from rateslib.calendars import dcf
 from rateslib.curves._parsers import _disc_maybe_from_curve, _disc_required_maybe_from_curve
 from rateslib.default import NoInput, _drb
 from rateslib.dual.utils import _dual_float
 from rateslib.periods.utils import _get_fx_and_base
+from rateslib.scheduling import dcf
 
 if TYPE_CHECKING:
     from rateslib.typing import (
