+++ python/rateslib/scheduling/adjuster.py
@@ -1,0 +1,45 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from rateslib.rs import Adjuster
+
+if TYPE_CHECKING:
+    pass
+
+_A = {  # Provides the map of all available string to Adjuster conversions.
+    "NONESETTLE": Adjuster.Actual(),
+    "NONE": Adjuster.Actual(),
+    "F": Adjuster.Following(),
+    "P": Adjuster.Previous(),
+    "MF": Adjuster.ModifiedFollowing(),
+    "MP": Adjuster.ModifiedPrevious(),
+    "FSETTLE": Adjuster.FollowingSettle(),
+    "PSETTLE": Adjuster.PreviousSettle(),
+    "MFSETTLE": Adjuster.ModifiedFollowingSettle(),
+    "MPSETTLE": Adjuster.ModifiedPreviousSettle(),
+}
+
+
+def _get_adjuster(adjuster: str | Adjuster) -> Adjuster:
+    """Convert a str such as 'F', 'MF' or '2B' or '5D' to an Adjuster."""
+    if isinstance(adjuster, Adjuster):
+        return adjuster
+    else:
+        adjuster = adjuster.upper()
+        if adjuster[-1] == "B":
+            return Adjuster.BusDaysLagSettle(int(adjuster[:-1]))
+        elif adjuster[-1] == "D":
+            return Adjuster.CalDaysLagSettle(int(adjuster[:-1]))
+        else:
+            return _A[adjuster]
+
+
+def _convert_to_adjuster(modifier: str, settlement: bool, mod_days: bool) -> Adjuster:
+    """Convert a legacy `modifier` to an Adjuster with additional options."""
+    modifier = modifier.upper()
+    if not mod_days and modifier[0] == "M":
+        modifier = modifier[1:]
+    if settlement:
+        modifier = modifier + "SETTLE"
+    return _get_adjuster(modifier)
