+++ python/tests/test_scheduling.py
@@ -6,9 +6,8 @@ from pandas import DataFrame, DatetimeIndex, date_range
 from pandas.testing import assert_index_equal
 from pandas.tseries.holiday import Holiday
 from rateslib import defaults
-from rateslib.calendars import Cal, create_calendar
 from rateslib.default import NoInput
-from rateslib.scheduling import Schedule
+from rateslib.scheduling import Cal, Schedule
 from rateslib.scheduling.scheduling import (
     _check_regular_swap,
     _check_unadjusted_regular_swap,
